+++ CHANGES.rst
@@ -9,6 +9,8 @@ ami
 associations
 ------------
 
+- Update rules so exclude dark files from associations [#4668]
+
 - Update association rules so that nodded observations procduce level 3 asn's [#4675]
 
 coron